Trying to figure out what's not getting deleted in participants directory.

+++ functions.php
@@ -4,14 +4,10 @@
       if ($dir = opendir($dir_name)) {
         while (($file_name = readdir($dir)) !== false) {
           $file_path = "$dir_name/$file_name";
-          echo "$file_path is current file.<br />";
-          if (!is_dir("$dir_name/$file")) {
-            if (is_dir($file_path)) {
-              clear_directory($file_path);
-            } else {
-              unlink($file_path);
-              echo "$file_path is cleared.<br />";
-            }
+          if (is_dir($file_path) and substr($file_name, 0, 1) != '.') {
+            clear_directory($file_path);
+          } else {
+            unlink($file_path);
           }
         }
         closedir($dir);
